+++ src/Eval.hs
@@ -1,7 +1,7 @@
 {-|
 Module      : Eval
 Description : Evalúa un término siguiendo la semántica big-step
-Copyright   : (c) Mauro Jaskelioff, Guido Martínez, Roman Castellarin, Sebastián Zimmermann, 2020.
+Copyright   : (c) Mauro Jaskelioff, Guido Martínez, Román Castellarin, Sebastián Zimmermann, 2020.
 License     : GPL-3
 Stability   : experimental
