+++ src/InlineExpansion.hs
@@ -2,7 +2,7 @@
 {-|
 Module      : InlineExpansion
 Description : Optimizacion de inline expansion
-Copyright   : (c) Roman Castellarin, Sebastián Zimmermann 2021.
+Copyright   : (c) Román Castellarin, Sebastián Zimmermann 2021.
 License     : GPL-3
 Stability   : experimental
 -}
